+++ common/src/main/java/net/mehvahdjukaar/vista/common/enderman/ScreenSpectatorView.java
@@ -1,6 +1,0 @@
-package net.mehvahdjukaar.vista.common.enderman;
-
-
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.phys.Vec2;
-